requirements/base.txt
# Core
Flask==2.3.3
Werkzeug==2.3.7

# API
flask-restx==1.2.0
marshmallow==3.20.1
webargs==8.2.0

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7
SQLAlchemy==2.0.19

# Authentication & Security
Flask-JWT-Extended==4.5.3
pyjwt==2.8.0
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4

# Caching & Rate Limiting
Flask-Caching==2.0.2
Flask-Limiter==3.3.1
redis==5.0.0

# Validation
pydantic==2.3.0
email-validator==2.1.0

# CORS
Flask-CORS==4.0.0

# Environment
python-dotenv==1.0.0

# Utilities
structlog==23.1.0
python-dateutil==2.8.2
python-magic==0.4.27
PyYAML==6.0